Fully resolved placeholder to avoid confusion (#57)
Updated the use of the _1 placeholder to use the fully resolved namespace to avoid confusing the compiler in the event that boost/function.hpp was included.

There are some headers that include boost function (e.g., OgreMeshManager.h) that cause a conflict when the placeholder is specified without its namespace. This keeps that from happening.